requirements.txt
# Flask and extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Marshmallow==0.14.0
Flask-CORS==4.0.0
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.23

# Development
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-flask==1.2.0